(core) move Job interface to taskExecutor
